Move chart component
