+++ src/components/JoinBlock.tsx
@@ -3,13 +3,12 @@ import clsx from "clsx";
 import dynamic from "next/dynamic";
 import Image from "next/image";
 import React from "react";
-const Plot = dynamic(() => import("react-plotly.js"), { ssr: false });
 
-import { useEffect, useState } from "react";
+const Chart = dynamic(() => import("@/components/Chart"), {
+  ssr: false,
+});
 
 import PrimaryButton from "@/components/buttons/PrimaryButton";
-
-import { fetchData } from "@/pages/api/chart";
 
 const IntroBlock = ({
   className,
@@ -18,21 +17,6 @@ const IntroBlock = ({
   className?: string;
   headDark?: boolean;
 }) => {
-  const [plot, setPlot] = useState({
-    data: [],
-    layout: {},
-  });
-  const fetchGraphData = async () => {
-    const response = await fetchData("graphs", "alternatives");
-    setPlot(response);
-  };
-  useEffect(() => {
-    fetchGraphData();
-  }, []);
-
-  if (!plot) {
-    return <></>;
-  }
   return (
     <div className={className}>
       <div
@@ -68,7 +52,7 @@ const IntroBlock = ({
             l'orientation des pratiques.
           </p>
           <div className="flex lg:justify-center min-h-[300px] overflow-y-auto">
-            <Plot data={plot.data} layout={plot.layout} />
+            <Chart id="alternatives" />
           </div>
         </div>
       </div>
